+++ src/App.js
@@ -2,9 +2,9 @@ import React from "react";
 import { Route, Switch } from "react-router-dom";
 import "./App.css";
 // * Component Imports
-import RoutesTree from "./RoutesTree";
-import Header from "./Components/HeaderComponents/Header";
-import NavBar from "./Components/NavBarComponents/NavBar";
+import RoutesTree from "./Components/RoutesTree";
+import Header from "./Components/SharedComponents/HeaderComponents/Header";
+import NavBar from "./Components/SharedComponents/NavBarComponents/NavBar";
 import Footer from "./Components/SharedComponents/Footer";
 
 function App() {
