--- src/Components/DanceStyles/BreakingInfo.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const BreakingInfo = () => {
-  return <div>Break styles bitch</div>;
-};
-
-export default BreakingInfo;
