--- src/Components/DanceStyles/HipHopInfo.jsx
@@ -1,49 +1,0 @@
-import React from "react";
-
-const HipHopInfo = () => {
-  return (
-    <div>
-      Hippity hoppity hop
-      <div>
-        Afrika bambaataa. Afrika Bambaataa (/ˌæfrɪkə bæmˈbɑːtə/; born Lance
-        Taylor,[2][3] April 17, 1957) is an American disc jockey, rapper,
-        songwriter and producer from the South Bronx, New York.[4][3] He is
-        notable for releasing a series of genre-defining electro tracks in the
-        1980s that influenced the development of hip hop culture.[5] Afrika
-        Bambaataa is one of the originators of breakbeat DJing and is
-        respectfully known as "The Godfather" and "Amen Ra of Hip Hop Kulture",
-        as well as the father of electro-funk.[1] Through his co-opting of the
-        street gang the Black Spades into the music and culture-oriented
-        Universal Zulu Nation, he has helped spread hip hop culture throughout
-        the world.[6] On May 6, 2016, Bambaataa left his position as head of The
-        Zulu Nation due to multiple child sexual abuse allegations dating as far
-        back as the 1970s.[7]
-      </div>
-      <div>
-        Cool Herc. Clive Campbell (born April 16, 1955), better known by his
-        stage name DJ Kool Herc, is a Jamaican DJ who is credited for
-        originating hip hop music in the Bronx, New York City, in the 1970s
-        through his "Back to School Jam",[4] hosted on August 11, 1973, at 1520
-        Sedgwick Avenue. After his younger sister, Cindy Campbell, became
-        inspired to earn extra cash for back-to-school clothes, she decided to
-        have her older brother, then 18 years old, play music for the
-        neighborhood in their apartment building.[5] Known as the "Founder of
-        Hip-Hop" and "Father of Hip-Hop", Campbell began playing hard funk
-        records of the sort typified by James Brown. Campbell began to isolate
-        the instrumental portion of the record which emphasized the drum
-        beat—the "break"—and switch from one break to another. Using the same
-        two-turntable set-up of disco DJs, he used two copies of the same record
-        to elongate the break. This breakbeat DJing, using funky drum solos,
-        formed the basis of hip hop music. Campbell's announcements and
-        exhortations to dancers helped lead to the syncopated, rhythmically
-        spoken accompaniment now known as rapping. He called the dancers
-        "break-boys" and "break-girls", or simply b-boys and b-girls. Campbell's
-        DJ style was quickly taken up by figures such as Afrika Bambaataa and
-        Grandmaster Flash. Unlike them, he never made the move into commercially
-        recorded hip hop in its earliest years.
-      </div>
-    </div>
-  );
-};
-
-export default HipHopInfo;
