--- src/Components/DanceStyles/HouseInfo.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const HouseInfo = () => {
-  return <div>House is JUMPIN</div>;
-};
-
-export default HouseInfo;
