--- src/Components/DanceStyles/LockingInfo.jsx
@@ -1,58 +1,0 @@
-import React from "react";
-
-const LockingInfo = () => {
-  return (
-    <div>
-      I lock a-lot
-      <div>
-        <h1>Locking</h1>
-        <p>
-          Locking is a style of funk dance, which is today also associated with
-          hip hop. The name is based on the concept of locking movements, which
-          basically means freezing from a fast movement and "locking" in a
-          certain position, holding that position for a short while and then
-          continuing at the same speed as before. It relies on fast and distinct
-          arm and hand movements combined with more relaxed hips and legs. The
-          movements are generally large and exaggerated, and often very rhythmic
-          and tightly synced with the music. Locking is performance oriented,
-          often interacting with the audience by smiling or giving them a high
-          five, and some moves are quite comical. Locking was originally danced
-          to traditional funk music, such as that produced or performed by James
-          Brown. Funk music is still commonly favored by locking dancers and
-          used by many competitions such as the locking division of Juste
-          Debout. Locking movements create a strong contrast towards the many
-          fast moves that are otherwise performed quite continuously, combined
-          with mime style performance and acting and other dancers. Locking
-          includes many acrobatics and physically demanding moves, such as
-          landing on one's knees and the split. These moves often require knee
-          protection.
-        </p>
-      </div>
-      <div>
-        <h1>History</h1>
-        The beginning of Locking can be traced to one man, Don Campbell. In the
-        late 1960s he put together several fad dances adding moves of his own
-        (known as the "Lock") when performing. The original lock was created by
-        accident: Campbell couldn't do a move called the 'Robot Shuffle' and
-        stopped at a particular point whilst moving his arms, creating a
-        'locking' effect.[1][2] He wasn't able to perform it fluently, for he
-        couldn't remember which step to take next. (Even the acting towards the
-        audience was spontaneous: when people started laughing at Don because of
-        his unfamiliar moves, he responded by pointing at them.) These halts
-        soon became popular as Don added them into his performances. The
-        resulting dance was called Campbellocking, which was later shortened to
-        Locking. In the early 1970s this set off a movement of Locking dance
-        groups, notably Campbell's group The Lockers. Other lockers Jimmy "Scoo
-        B Doo" Foster, Greggory "Campbellock Jr." Pope, Tony "GoGo" Lewis, Fred
-        "Mr Penguin" Berry (a.k.a. Rerun), Leo "Fluky Luke" Williamson, Damita
-        Jo Freeman and others also helped set the foundation for the locking
-        dance and clothes style. Clothes style can consist of loud striped
-        socks, pegged pants that stopped at the knees, bright colorful satin
-        shirts with big collars, big colorful bow ties, gigantic Apple Boy hats,
-        and white gloves.
-      </div>
-    </div>
-  );
-};
-
-export default LockingInfo;
