--- src/Components/DanceStyles/PoppingInfo.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const PoppingInfo = () => {
-  return <div>Popping dem BEATS</div>;
-};
-
-export default PoppingInfo;
