--- src/Components/DanceStyles/VogueingInfo.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const VogueingInfo = () => {
-  return <div>Vogue away hunny</div>;
-};
-
-export default VogueingInfo;
