--- src/Components/DanceStyles/WaackingInfo.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const WaackingInfo = () => {
-  return <div>Waackkkk AAATAAAACKKK!</div>;
-};
-
-export default WaackingInfo;
