--- src/Pages/EventPage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const EventPage = (props) => {
-  return <div>Event Page Container</div>;
-};
-
-export default EventPage;
